assistant/requirements.txt
fastapi
pydantic-ai
logfire[fastapi]
sse-starlette
uvicorn
openai
tavily-python
